Refactor encoding.Packer to have Mercury control its own interface and function implementations
